fixing some struct locations for idiomatic imports
