+++ src/websocket/server/session.rs
@@ -1,5 +1,5 @@
 use super::Message;
-use crate::game::elements::point::Point;
+use crate::game::elements::Point;
 use async_trait::async_trait;
 use tracing::info;
